fix pettingzoo(wzl):update config and pettingzoo env
